+++ test-app/src/androidTest/java/com/linkedin/parser/test/junit4/kotlin/InterfaceWithDefaultMethod.kt
@@ -1,0 +1,9 @@
+package com.linkedin.parser.test.junit4.kotlin
+
+import org.junit.Test
+
+interface InterfaceWithDefaultMethod {
+    @Test
+    fun testMethodShouldNotBeReported() {
+    }
+}
